requirements.txt
Django==1.8
django-admin-bootstrapped==2.5.6
django-ckeditor-updated==4.4.4
django-classy-tags==0.6.2
django-jet==0.1.3
django-mysql==1.0.1
django-polymorphic==0.7.2
-e git+https://github.com/divio/django-shop.git@a74e38369de40b9e5f481f6ac9dda6d5eb161da0#egg=django_shop-master
django-taggit==0.17.3
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
jsonfield==1.0.3
mysqlclient==1.3.7
Pillow==3.0.0
six==1.10.0
wheel==0.24.0
